requirements.txt
# fastapi[all]
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==0.21.1